sales_admin/requirements.txt
alembic==1.0.11
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.14
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
